finished most of the commands in the batch api, need to talk to dani

+++ carto/sql_api.py
@@ -1,5 +1,7 @@
 import requests
 import time
+from carto import NoAuthClient
+from .core import CartoException
 
 SQL_API_URL = '{api_version}/sql'
 SQL_BATCH_API_URL = '{api_version}/sql/job/'
@@ -45,39 +47,43 @@ class BatchSQLClient(object):
 
     def __init__(self, client, api_version='v2'):
         self.client = client
-        self.api_key = self.client.api_key
         self.api_url = SQL_BATCH_API_URL.format(api_version=api_version)
+        if not isinstance(self.client, NoAuthClient):
+            self.api_key = self.client.api_key
 
     def create(self, sql_query):
-        #raise cartoexception
+        if isinstance(self.client, NoAuthClient):
+            raise CartoException("The client must be authenticated with an API key to create a batch sql job")
         header = {'content-type': 'application/json'}
         data = requests.post('https://rsharan.cartodb.com/api/' + self.api_url + "?api_key=" + self.api_key, json={"query": sql_query}, headers=header)
-        #print(data.json())
-        return data.json()['job_id']
+        return data.json()
 
     def read(self, job_id):
-        #split into 2 cases
         header = {'content-type': 'application/json'}
-        data = requests.get('https://rsharan.cartodb.com/api/' + self.api_url + job_id + "?api_key=" + self.api_key, headers=header)
+        if isinstance(self.client, NoAuthClient):
+            data = requests.get('https://rsharan.cartodb.com/api/' + self.api_url + job_id, headers=header)
+        else:
+            data = requests.get('https://rsharan.cartodb.com/api/' + self.api_url + job_id + "?api_key=" + self.api_key, headers=header)
         #print(data.json())
         return data.json()
 
     def update(self, job_id, sql_query):
-        #raise cartoexception
+        if isinstance(self.client, NoAuthClient):
+            raise CartoException("The client must be authenticated with an API key to update a batch sql job")        
         header = {'content-type': 'application/json'}
         data = requests.put('https://rsharan.cartodb.com/api/' + self.api_url + job_id + "?api_key=" + self.api_key, json={"query": sql_query}, headers=header)
-        print(data.json())
+        #print(data.json())
         return data.json()
 
     def cancel(self, job_id):
-        #raise cartoexception
+        if isinstance(self.client, NoAuthClient):
+            raise CartoException("The client must be authenticated with an API key to update a batch sql job")        
         confirmation = requests.delete('https://rsharan.cartodb.com/api/' + self.api_url + job_id + "?api_key=" + self.api_key)
+        print(confirmation.json())
         return confirmation.json()['status']
 
 
 #for batch requests, do testing for both lists and string lists
-#use large dataset
-
 
 class BatchSQLManager(object):
     def __init__(self, client, api_version='v2'):
@@ -86,7 +92,8 @@ class BatchSQLManager(object):
         self.api_url = SQL_BATCH_API_URL.format(api_version=api_version)
 
     def all(self):
-        #raise cartoexception
+        if isinstance(self.client, NoAuthClient):
+            raise CartoException("The client must be authenticated with an API key to list the sql jobs")
         data = requests.get('https://rsharan.cartodb.com/api/' + self.api_url + "?api_key=" + self.api_key)
         return data.json()
